Post VIE cleanup changes
